id long

+++ gestion-services/src/main/java/com/alma/enseignants/CasSpecial.java
@@ -2,7 +2,7 @@ package com.alma.enseignants;
 
 public class CasSpecial extends Intervention{
 
-	public CasSpecial(int volume, Service service, int id) {
+	public CasSpecial(int volume, Service service, long id) {
 		super(volume, service, id);
 	}
