+++ gestion-services/src/main/java/com/alma/enseignants/InterventionDep.java
@@ -2,7 +2,7 @@ package com.alma.enseignants;
 
 public class InterventionDep extends Intervention{
 
-	public InterventionDep(int volume, Service service, int id) {
+	public InterventionDep(int volume, Service service, long id) {
 		super(volume, service, id);
 	}
 }
