+++ gestion-services/src/main/java/com/alma/enseignants/InterventionExt.java
@@ -2,7 +2,7 @@ package com.alma.enseignants;
 
 public class InterventionExt extends Intervention{
 
-	public InterventionExt(int volume, Service service, int id) {
+	public InterventionExt(int volume, Service service, long id) {
 		super(volume, service, id);
 	}
